requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
facebook-sdk==3.0.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
requests==2.19.1
SQLAlchemy==1.2.11
urllib3==1.23
Werkzeug==0.14.1
